==> 03-PREPROCESSING-FOR-FEATURES.ipynb <==
import numpy as np
import pandas as pd


daily_data = pd.DataFrame(pd.read_csv("consumption_daily.csv",header=0, index_col=0, parse_dates=True, squeeze=True))


# Adding seasonality features

## Set weekends and holidays to 1, otherwise 0
daily_data['Atypical_Day'] = np.zeros(len(daily_data))
daily_data["Religious Holidays"] = np.zeros(len(daily_data))

# Weekends 5 saturday, 6 sunday
daily_data['Atypical_Day'][(daily_data.index.dayofweek==5)|(daily_data.index.dayofweek==6)] = 1

# Holidays
holidays = [str(x)+ a for a in ["-01-01","-04-23","-05-01","-05-19","-07-15","-08-30","-10-29"] for x in range(2016,2020)]
bayramlar = [
 '2016-07-04',
 '2016-07-05',
 '2016-07-06',
 '2016-07-07',
 '2016-09-11',
 '2016-09-12',
 '2016-09-13',
 '2016-09-14',
 '2016-09-15',
 '2016-12-31',
 '2017-06-24',
 '2017-06-25',
 '2017-06-26',
 '2017-06-27',
 '2017-08-31',
 '2017-09-01',
 '2017-09-02',
 '2017-09-03',
 '2017-09-04',
 '2017-12-31',
 '2018-06-14',
 '2018-06-15',
 '2018-06-16',
 '2018-06-17',
 '2018-08-20',
 '2018-08-21',
 '2018-08-22',
 '2018-08-23',
 '2018-08-24',
 '2018-12-31',
 '2019-06-04',
 '2019-06-05',
 '2019-06-06',
 '2019-06-07',
 '2019-08-10',
 '2019-08-11',
 '2019-08-12',
 '2019-08-13',
 '2019-08-14',
 '2019-12-31'
]

for gun,j in daily_data.iterrows():    
    if str(gun)[:10] in bayramlar:
        daily_data["Religious Holidays"].loc[gun] =1
        

for gun,j in daily_data.iterrows():    
    if str(gun)[:10] in holidays:
        daily_data["Atypical_Day"].loc[gun] =1
 

daily_data.head(5)

# Saving the data with seasonality features to create different window sizes.
daily_data.to_csv("daily_data_seasonality.csv",index = True)

# Creating Window Features (7, 14, 21, and 28)

# Add historic usage to each X vector
daily_data = pd.DataFrame(pd.read_csv("daily_data_seasonality.csv",header=0, index_col=0, parse_dates=True, squeeze=True))
# Set number of days prediction is in advance
n_days_advance = 1

# Set number of historic days used
n_days_window = 7

# Adding corresponding columns for windows
for k in range(n_days_advance,n_days_advance+n_days_window):
    daily_data['Consumption_t-%i'% k] = np.zeros(len(daily_data['Consumption']))

# Adding historic consumption to corresponding columns
for i in range(n_days_advance+n_days_window,len(daily_data['Consumption'])):
    for j in range(n_days_advance,n_days_advance+n_days_window):
        daily_data['Consumption_t-%i'% j][i] = daily_data['Consumption'][i-j]


# Eliminating the data which has no window size.
daily_data = daily_data.iloc[n_days_advance+n_days_window:]
daily_data.to_csv("daily_data_seasonality_7.csv",index = True)
daily_data.head(3)

### Applying same steps for different window size according to statistical analysis (14, 21, and 28).

# Add historic usage to each X vector
daily_data = pd.DataFrame(pd.read_csv("daily_data_seasonality.csv",header=0, index_col=0, parse_dates=True, squeeze=True))
# Set number of days prediction is in advance
n_days_advance = 1

# Set number of historic days used
n_days_window = 14

   
for k in range(n_days_advance,n_days_advance+n_days_window):
    daily_data['Consumption_t-%i'% k] = np.zeros(len(daily_data['Consumption']))

    
for i in range(n_days_advance+n_days_window,len(daily_data['Consumption'])):
    for j in range(n_days_advance,n_days_advance+n_days_window):
        daily_data['Consumption_t-%i'% j][i] = daily_data['Consumption'][i-j]

daily_data = daily_data.iloc[n_days_advance+n_days_window:]
daily_data.to_csv("daily_data_seasonality_14.csv",index = True)
daily_data.head(3)

# Add historic usage to each X vector
daily_data = pd.DataFrame(pd.read_csv("daily_data_seasonality.csv",header=0, index_col=0, parse_dates=True, squeeze=True))
# Set number of days prediction is in advance
n_days_advance = 1

# Set number of historic days used
n_days_window = 21


for k in range(n_days_advance,n_days_advance+n_days_window):
    
    daily_data['Consumption_t-%i'% k] = np.zeros(len(daily_data['Consumption']))

    
    
for i in range(n_days_advance+n_days_window,len(daily_data['Consumption'])):
    
    for j in range(n_days_advance,n_days_advance+n_days_window):
        
        daily_data['Consumption_t-%i'% j][i] = daily_data['Consumption'][i-j]
      

daily_data = daily_data.iloc[n_days_advance+n_days_window:]
daily_data.to_csv("daily_data_seasonality_21.csv",index = True)
daily_data.head(3)

# Add historic usage to each X vector
daily_data = pd.DataFrame(pd.read_csv("daily_data_seasonality.csv",header=0, index_col=0, parse_dates=True, squeeze=True))
# Set number of days prediction is in advance
n_days_advance = 1

# Set number of historic days used
n_days_window = 28


for k in range(n_days_advance,n_days_advance+n_days_window):
    
    daily_data['Consumption_t-%i'% k] = np.zeros(len(daily_data['Consumption']))

    
    
for i in range(n_days_advance+n_days_window,len(daily_data['Consumption'])):
    
    for j in range(n_days_advance,n_days_advance+n_days_window):
        
        daily_data['Consumption_t-%i'% j][i] = daily_data['Consumption'][i-j]
      

daily_data = daily_data.iloc[n_days_advance+n_days_window:]
daily_data.to_csv("daily_data_seasonality_28.csv",index = True)
daily_data.head(3)

# 4 datasets are created.
## 7-days window, 14-days window, 21-days window, and 28-days window.

daily_data.head()
