==> 04-ARIMA-MODELS.ipynb <==
# Seasonal Autoregressive Integrated Moving Averages Model with Exogenous Variables

!pip install pmdarima
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import pmdarima as pm
import statsmodels.api as sm

daily_data = pd.DataFrame(pd.read_csv("consumption_daily.csv",header=0, index_col=0, parse_dates=True, squeeze=True))
daily_data_train= pd.DataFrame(daily_data["2016":"2018"],copy=True)
daily_data_test = pd.DataFrame(daily_data["2019"],copy=True)

tatil_data = pd.DataFrame(np.zeros(shape=(daily_data.count(0)[0],3)),index=daily_data.index,columns=["Tatil","Bayram","Haftasonu"])
tatil_gunleri = [str(x)+ a for a in ["-01-01","-04-23","-05-01","-05-19","-07-15","-08-30","-10-29"] for x in range(2016,2020)]
tatil_data['Haftasonu'][(daily_data.index.dayofweek==5)|(daily_data.index.dayofweek==6)] = 1
for gun in tatil_gunleri:
    tatil_data["Tatil"].loc[gun] = 1

bayramlar = ['2016-07-04', '2016-07-05', '2016-07-06', '2016-07-07',
             '2016-09-11', '2016-09-12', '2016-09-13', '2016-09-14','2016-09-15',
             '2017-06-24', '2017-06-25', '2017-06-26', '2017-06-27',
             '2017-08-31', '2017-09-01', '2017-09-02', '2017-09-03', '2017-09-04',
             '2018-06-14', '2018-06-15', '2018-06-16', '2018-06-17',
             '2018-08-20', '2018-08-21', '2018-08-22', '2018-08-23', '2018-08-24', 
             '2019-06-04', '2019-06-05', '2019-06-06', '2019-06-07', 
             '2019-08-10', '2019-08-11', '2019-08-12', '2019-08-13', '2019-08-14']
#Setting the corresponding dates to 1.
for gun in tatil_gunleri:
    tatil_data["Tatil"].loc[gun] = 1
for gun in bayramlar:
    tatil_data["Bayram"].loc[gun] =1

tatil_data.head()

## Searching Models Stepwise through SARIMAX(p,d,q)(P,D,Q)(7) parameters
## Building Models with different exogenous features

model_noExog = pm.auto_arima(daily_data[:"2018"],
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)



bayram_df = pd.DataFrame(tatil_data[:"2018"]["Bayram"])
bayram_df.index = daily_data[:"2018"].index

model_onlyBayram = pm.auto_arima(daily_data["Consumption"][:"2018"], exogenous=bayram_df,
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)

tatil_df = pd.DataFrame(tatil_data[:"2018"]["Tatil"])
tatil_df.index = daily_data[:"2018"].index

model_onlyTatil = pm.auto_arima(daily_data["Consumption"][:"2018"], exogenous=tatil_df,
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)

haftasonu_df = pd.DataFrame(tatil_data[:"2018"]["Haftasonu"])
haftasonu_df.index = daily_data[:"2018"].index

model_onlyHaftasonu = pm.auto_arima(daily_data["Consumption"][:"2018"], exogenous=haftasonu_df,
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)

bayram_tatil_df = pd.DataFrame(tatil_data[:"2018"][["Bayram","Tatil"]])
bayram_tatil_df.index = daily_data[:"2018"].index


model_noHaftasonu = pm.auto_arima(daily_data["Consumption"][:"2018"], exogenous=bayram_tatil_df,
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)

haftasonu_tatil_df = pd.DataFrame(tatil_data[:"2018"][["Tatil","Haftasonu"]])
haftasonu_tatil_df.index = daily_data[:"2018"].index


model_noBayram = pm.auto_arima(daily_data["Consumption"][:"2018"], exogenous=haftasonu_tatil_df,
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)

bayram_hs_df = pd.DataFrame(tatil_data[:"2018"][["Bayram","Haftasonu"]])
bayram_hs_df.index = daily_data[:"2018"].index


model_noTatil = pm.auto_arima(daily_data["Consumption"][:"2018"], exogenous=bayram_hs_df,
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)

model_all = pm.auto_arima(daily_data["Consumption"][:"2018"], exogenous=tatil_data[:"2018"],
                     max_d = 2, max_p=10, max_q=10, max_P=10, max_Q=10, max_D = 10, seasonal=True,
                     m=7,
                     stepwise=True, suppress_warnings=True, with_intercept=True,
                     error_action='ignore',maxiter = 50)


#Models Dictionary
models = {"No Exogenous Model":model_noExog,
          "Only Bayram Exog. Model":model_onlyBayram,
          "Only Tatil Exog. Model":model_onlyTatil,
          "Only Haftasonu Exog. Model":model_onlyHaftasonu,
          "No Bayram Model":model_noBayram,
          "No Haftasonu Model":model_noHaftasonu,
          "No Tatil Model":model_noTatil,
          "All Including Model":model_all}


model_all.summary()

### Fitting models

model_noExog_fit = model_noExog.fit(daily_data["Consumption"][:"2018"])

model_onlyBayram_fit = model_onlyBayram.fit(daily_data["Consumption"][:"2018"],exogenous=bayram_df)
model_onlyTatil_fit = model_onlyBayram.fit(daily_data["Consumption"][:"2018"],exogenous=tatil_df)
model_onlyHaftasonu_fit = model_onlyBayram.fit(daily_data["Consumption"][:"2018"],exogenous=haftasonu_df)

model_noHaftasonu_fit = model_noHaftasonu.fit(daily_data["Consumption"][:"2018"],exogenous=bayram_tatil_df)
model_noTatil_fit = model_noTatil.fit(daily_data["Consumption"][:"2018"],exogenous=bayram_hs_df)
model_noBayram_fit = model_noTatil.fit(daily_data["Consumption"][:"2018"],exogenous=haftasonu_tatil_df)

model_all_fit = model_all.fit(daily_data["Consumption"][:"2018"],exogenous=tatil_data[:"2018"])


models_fit = {"No Exogenous Model":model_noExog_fit,
          "Only Bayram Exog. Model":model_onlyBayram_fit,
          "Only Tatil Exog. Model":model_onlyTatil_fit,
          "Only Haftasonu Exog. Model":model_onlyHaftasonu_fit,
          "No Bayram Exog. Model":model_noBayram_fit,
          "No Haftasonu Model":model_noHaftasonu_fit,
          "No Tatil Model":model_noTatil_fit,
          "All Including Model":model_all_fit}
#fitting the daily training data

def forecast_update(testData,model,tatilData):
    forecast = pd.DataFrame(np.zeros(testData.index.size),index=testData.index,columns=["Consumption"])
    for i,j in testData.iteritems() :
        forecast.loc[i]= model.predict(exogenous=np.array(tatilData.loc[i]).reshape(1,-1),n_periods=1)[0]
        model.update(testData.loc[i],exogenous=np.array(tatilData.loc[i]).reshape(1,-1))
    return forecast



type(daily_data) is pd.DataFrame

## EXOGENOUS VARIABLES FOR TEST

bayram_b_hs_df = pd.DataFrame(tatil_data[["Bayram","Haftasonu"]]["2019-01-01":"2019-12-31"])
bayram_b_hs_df.index = daily_data["2019"].index

bayram_b_t_df = pd.DataFrame(tatil_data[["Bayram","Tatil"]]["2019-01-01":"2019-12-31"])
bayram_b_t_df.index = daily_data["2019"].index

bayram_t_hs_df = pd.DataFrame(tatil_data[["Tatil","Haftasonu"]]["2019-01-01":"2019-12-31"])
bayram_t_hs_df.index = daily_data["2019"].index

bayram_b_df = pd.DataFrame(tatil_data["Bayram"]["2019-01-01":"2019-12-31"])
bayram_b_df.index = daily_data["2019"].index

bayram_t_df = pd.DataFrame(tatil_data["Tatil"]["2019-01-01":"2019-12-31"])
bayram_t_df.index = daily_data["2019"].index

bayram_hs_df = pd.DataFrame(tatil_data["Haftasonu"]["2019-01-01":"2019-12-31"])
bayram_hs_df.index = daily_data["2019"].index


#Forecasts with one step at a time and updating the model

forecast_onlyBayram = forecast_update(daily_data["Consumption"]["2019-01-01":"2019-12-31"],model_onlyBayram_fit,bayram_b_df)
forecast_onlyTatil = forecast_update(daily_data["Consumption"]["2019-01-01":"2019-12-31"],model_onlyTatil_fit,bayram_t_df)
forecast_onlyHaftasonu = forecast_update(daily_data["Consumption"]["2019-01-01":"2019-12-31"],model_onlyHaftasonu_fit,bayram_hs_df)


forecast_noBayram = forecast_update(daily_data["Consumption"]["2019-01-01":"2019-12-31"],model_noBayram_fit,bayram_t_hs_df)
forecast_noHaftasonu = forecast_update(daily_data["Consumption"]["2019-01-01":"2019-12-31"],model_noHaftasonu_fit,bayram_b_t_df)
forecast_noTatil = forecast_update(daily_data["Consumption"]["2019-01-01":"2019-12-31"],model_noTatil_fit,bayram_b_hs_df)

forecast_all = forecast_update(daily_data["Consumption"]["2019-01-01":"2019-12-31"],model_all_fit,tatil_data["2019-01-01":"2019-12-31"])
forecast_noExog = pd.DataFrame(np.zeros(daily_data["Consumption"]["2019-01-01":"2019-12-31"].index.size),index=daily_data["Consumption"]["2019-01-01":"2019-12-31"].index,columns=["Consumption"])

for i,j in daily_data["Consumption"]["2019-01-01":"2019-12-31"].iteritems() :
      forecast_noExog.loc[i]= model_noExog_fit.predict(n_periods=1)[0]
      model_noExog_fit.update(daily_data["Consumption"]["2019-01-01":"2019-12-31"].loc[i])

models_forecast = {"No Exogenous Model":forecast_noExog,
         "Only Bayram Exog. Model":forecast_onlyBayram,
         "Only Tatil Exog. Model":forecast_onlyTatil,
         "Only Haftasonu Exog. Model":forecast_onlyHaftasonu,
         "No Bayram Exog. Model":forecast_noBayram,                  
         "No Haftasonu Model":forecast_noHaftasonu,
        "No Tatil Model":forecast_noTatil,
         "All Including Model":forecast_all}

#Correlations of forecasts with actual data.
print("Correlations of forecasts with actual data.")
for model in models_forecast:
    correlation = daily_data_test["2019-01-01":"2019-12-31"].Consumption.corr(models_forecast[model].Consumption)
    print(model,":",correlation)

mapes = {}
for model in models_forecast :
    print("\n")
    print("-"*10,model,"-"*10)
    errors = (daily_data_test["2019-01-01":"2019-12-31"]-models_forecast[model])
    errors_percentage = ((errors/daily_data_test["2019-01-01":"2019-12-31"])*100)
    MPE = errors_percentage.mean()[0]
    MAPE = abs(errors_percentage).mean()[0]
    mapes[model] = MAPE 

print("Best Mape is: ",min(mapes, key=mapes.get))

model_noBayram.summary()

# Saving the best models predicts

index_column = pd.date_range(start ='2019-1-1', end = '2019-12-31', freq ='D')
best_pred = pd.DataFrame(models_forecast[min(mapes, key=mapes.get)],index=index_column)
best_pred.to_csv("the_best_arima_pred.csv",index = True)


mapes_df = pd.DataFrame.from_dict(mapes, orient='index')
mapes_df.to_csv("mapes_arima.csv",index=True)
